pyproject.toml
[project]
name = "dedup"
version = "0.1.0"
description = "CLI tool for finding duplicate and similar images across NAS backups"
readme = "README.md"
authors = [
    { name = "Jason Hussey", email = "jphussey05@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15.0",
    "rich>=14.0.0",
    "Pillow>=11.0.0",
    "pillow-heif>=1.3.0",
    "imagehash>=4.3.0",
]

[project.optional-dependencies]
cnn = [
    "sentence-transformers>=3.0.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "faiss-cpu>=1.9.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.8.0",
]
all = ["dedup[cnn,dev]"]

[project.scripts]
dedup = "dedup.cli:app"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_model: marks tests that require downloading ML models",
]
